% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dois}
\alias{dois}
\title{DOIs dataset}
\format{
A data frame with columns: country, year, type, label, id.
}
\source{
Data curated by the Development Engagement Lab.
}
\usage{
dois
}
\description{
A dataset containing DOIs for various datasets.
}
\keyword{datasets}
