% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_del_data.R
\name{get_del_data}
\alias{get_del_data}
\title{DELdata: An R package to download Development Engagement Lab datasets}
\usage{
get_del_data(country = NULL, year = NULL, type = NULL)
}
\arguments{
\item{country}{A vector of countries. Options include "France", "Germany", "Great Britain", and "United States".}

\item{year}{A vector containing one or several years. Acceptable values are from 2019 to 2024.}

\item{type}{A vector containing the data type. Options include "Panel", "Tracker", and "Sandbox".}
}
\value{
A list of data frames containing the requested datasets.
}
\description{
The main function of this package is \code{get_del_data()}, which allows you to download DEL's data based on the country, type, and years.
}
\examples{
get_del_data(country = 'France', year = '2020', type = 'Panel')
}
